Add standalone leaderboard endpoint
