+++ src/routes/quiz.ts
@@ -3,10 +3,9 @@ import { PromptBuilder } from '@/services/PromptBuilder'
 
 // Services
 import { SecureToolHandler } from '@/services/SecureToolHandler'
-import { ScoringService } from '@/services/ScoringService'
 
 // Middleware
-import { ValidationMiddleware, schemas, querySchemas } from '@/middleware/validation'
+import { ValidationMiddleware, schemas } from '@/middleware/validation'
 import { AuthMiddleware, getAuthenticatedUser } from '@/middleware/auth'
 
 // Types
@@ -75,66 +74,6 @@ router.post(
     }
   }
 )
-
-// Leaderboard endpoint
-router.get(
-  '/leaderboard',
-  ValidationMiddleware.validateQuery(querySchemas.leaderboard),
-  async (c: AppContext) => {
-    try {
-      const db = c.get('db')
-      const validatedQuery = c.get('validatedQuery') as {
-        limit: number
-        offset: number
-        period: 'all' | 'today' | 'week' | 'month'
-      }
-
-      const scoringService = new ScoringService(db)
-      const fullLeaderboard = await scoringService.getLeaderboard(validatedQuery.limit + validatedQuery.offset)
-      
-      // Apply pagination
-      const paginatedResults = fullLeaderboard.slice(validatedQuery.offset, validatedQuery.offset + validatedQuery.limit)
-
-      return c.json({
-        success: true,
-        data: {
-          entries: paginatedResults,
-          pagination: {
-            total: fullLeaderboard.length,
-            limit: validatedQuery.limit,
-            offset: validatedQuery.offset,
-            hasMore: validatedQuery.offset + validatedQuery.limit < fullLeaderboard.length
-          }
-        },
-        timestamp: new Date().toISOString()
-      })
-
-    } catch (error) {
-      logger.error('Leaderboard fetch failed:', error)
-      
-      if (error instanceof AppError) {
-        return c.json({
-          success: false,
-          error: {
-            code: error.code,
-            message: error.message
-          },
-          timestamp: new Date().toISOString()
-        }, error.statusCode as any)
-      }
-
-      return c.json({
-        success: false,
-        error: {
-          code: ErrorCode.INTERNAL_SERVER_ERROR,
-          message: 'Leaderboard fetch failed'
-        },
-        timestamp: new Date().toISOString()
-      }, 500)
-    }
-  }
-)
-
 // Export router
 // Ephemeral token endpoint for OpenAI Realtime API
 router.post('/realtime/ephemeral-token', async (c: AppContext) => {
